requirements.txt
Django==4.1.9
pymongo==3.12.3
django-crispy-forms==2.0
django-bootstrap4==23.1
djongo==1.3.6
pytz==2023.3
python_dotenv==0.21.0
